+++ internal/query/session_test.go
@@ -1,0 +1,56 @@
+package query
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Query"
+	"go.uber.org/mock/gomock"
+	grpcCodes "google.golang.org/grpc/codes"
+	grpcStatus "google.golang.org/grpc/status"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xtest"
+	"github.com/ydb-platform/ydb-go-sdk/v3/query"
+)
+
+func TestBegin(t *testing.T) {
+	t.Run("HappyWay", func(t *testing.T) {
+		ctx := xtest.Context(t)
+		ctrl := gomock.NewController(t)
+		service := NewMockQueryServiceClient(ctrl)
+		service.EXPECT().BeginTransaction(gomock.Any(), gomock.Any()).Return(&Ydb_Query.BeginTransactionResponse{
+			Status: Ydb.StatusIds_SUCCESS,
+			TxMeta: &Ydb_Query.TransactionMeta{
+				Id: "123",
+			},
+		}, nil)
+		t.Log("begin")
+		tx, err := begin(ctx, service, "123", query.TxSettings())
+		require.NoError(t, err)
+		require.Equal(t, "123", tx.id)
+	})
+	t.Run("TransportError", func(t *testing.T) {
+		ctx := xtest.Context(t)
+		ctrl := gomock.NewController(t)
+		service := NewMockQueryServiceClient(ctrl)
+		service.EXPECT().BeginTransaction(gomock.Any(), gomock.Any()).Return(nil, grpcStatus.Error(grpcCodes.Unavailable, ""))
+		t.Log("begin")
+		_, err := begin(ctx, service, "123", query.TxSettings())
+		require.Error(t, err)
+		require.True(t, xerrors.IsTransportError(err, grpcCodes.Unavailable))
+	})
+	t.Run("OperationError", func(t *testing.T) {
+		ctx := xtest.Context(t)
+		ctrl := gomock.NewController(t)
+		service := NewMockQueryServiceClient(ctrl)
+		service.EXPECT().BeginTransaction(gomock.Any(), gomock.Any()).Return(&Ydb_Query.BeginTransactionResponse{
+			Status: Ydb.StatusIds_UNAVAILABLE,
+		}, nil)
+		t.Log("begin")
+		_, err := begin(ctx, service, "123", query.TxSettings())
+		require.Error(t, err)
+		require.True(t, xerrors.IsOperationError(err, Ydb.StatusIds_UNAVAILABLE))
+	})
+}